Crate slog_conf

Highly customizable runtime-configuration for slog-rs/slog with opinionated defaults.

Overview

This crate is roughly divided into two parts:

De-/serialization of configuration files.

Construction of a Drain (or Logger) from a configuration object.

Both aspects can be configured, all defaults can be completely replaced.

Most of the functionality is centered around the Config trait and trait-objects of that type. This trait describes a configuration with which a logger (or drain) can be built.

De-/Serialization

Serialization of configurations is straight-forward except for a small detail. Configurations should always be serialized as trait-object, otherwise their type tag will not be included during serialization and thus deserialization will fail.

Deserialization of Box<Config> can be acheived by use of the deserialize function of a Deserializers-registry. A default registry is provided by the deserializers-method. This default registry will be used if Box<Config> is directly deserialized.

Custom deserialization can, for example, be implemented with a newtype-wrapper for Box<Config> and a custom registry.

Building a Logger

Constructing a logger from a Config trait-object can be done via a Factories registry. A default registry is provided via the factories-method. build is a convenience-method using this default registry to build a Drain.

Customizable Features for Compile-Time Configuration

The configuration types and default factories supported by this crate can be found in the ty module and can be configured via the feature-set of this crate. For each supported type exists is a corresponding feature with the same name enabling support for said type. By default, all types are enabled.

Re-exports

pub use ty::null::Config as NullConfig;

pub use ty::null::Factory as NullFactory;

pub use ty::plain::Config as PlainConfig;

pub use ty::plain::Factory as PlainFactory;

pub use ty::term::Config as TermConfig;

pub use ty::term::Factory as TermFactory;

pub use ty::json::Config as JsonConfig;

pub use ty::json::Factory as JsonFactory;

Modules

common
Common configuration types.

erased
Utilities for type-erased de-/serialization.

ty
Default configurations and factories.

Structs

Deserializers
A registry for DeserializeSeed implementations for deserialization of Config trait-objects.

Factories
A registry for factories.

Enums

Error
An error that can occur when building a logger.

Constants

SUPPORTED_TYPES
All logger types supported by this crate.

TYPE_KEY
The name of the field containing the type of a serialized logger configuration.

Traits

Config
A configuration describing how a logger should be created.

Factory
A factory that can build a target from a configuration.

Functions

build
Builds a Drain from the given Config using the default factories.

deserializers
Returns a reference to the default deserializer-stub registry.

factories
Returns a reference to the default Drain factories.