Struct slog_conf::Deserializers

pub struct Deserializers { /* fields omitted */ }

A registry for DeserializeSeed implementations for deserialization of Config trait-objects.

This registry allows for deserialization of Config trait-objects based on their type tag by providing a dynamically dispatched implementation of DeserializeSeed.

Methods

impl Deserializers

pub fn empty() -> Self

Creates a new, empty registry.

pub fn register<F>(
&mut self,
tag: &'static str,
seed: F
) -> Option<BoxFnSeed<Box<Config>>> where
F: FnSeed<Box<Config>> + Sync + 'static,

Registers a closure as DeserializeSeed for the specified type-tag.

Thre registered deserializer will be used for deserializing a Config object with the specified value as type.

Returns the DeserializeSeed previously associated with the specified type-tag, or None if no such entry existed.

This method is a convenience-wrapper around insert.