Trait slog_conf::Config

pub trait Config: Serialize + Debug + 'static {
    fn ty(&self) -> &'static str;

    fn type_id(&self) -> TypeId { ... }
}

A configuration describing how a logger should be created.

Required Methods

fn ty(&self) -> &'static str

The type-tag of this configuration.

Provided Methods

fn type_id(&self) -> TypeId

The type-id of the configuration implementation.

Warning

You should not implement this method manually.

Methods

impl Config

pub fn is<T: Config>(&self) -> bool

Returns true if the actual type of this trait-object is the same as T.

pub fn downcast_ref<T: Config>(&self) -> Option<&T>

Returns some reference to the actual value of this trait object if it is of type T, or None if it is not.

pub fn downcast_mut<T: Config>(&mut self) -> Option<&mut T>

Returns some mutable reference to the actual value of this trait object if it is of type T, or None if it is not.

pub fn apply_ref<T, R, F>(&self, visitor: F) -> Option<R> where
T: Config + Sized,
F: FnOnce(&T) -> R,

Apply the given reference-visitor to this configuration.

The visitor will only be executed if the type of this configuration is T, and the result provided by the visitor will be returned. Otherwise None will be returned.

pub fn apply_mut<T, R, F>(&mut self, visitor: F) -> Option<R> where
T: Config + Sized,
F: FnOnce(&mut T) -> R,

Apply the given mutable-reference-visitor to this configuration.

The visitor will only be executed if the type of this configuration is T, and the result provided by the visitor will be returned. Otherwise None will be returned.